Feature/custom map layers (#451)
* refactor map container

* refactor map buttons

* refactor choose on map

* refactor maps folder

* refactor map utils

* refactor names

* adding custom layers

* adding dummy busses

* adding settings custom layer config

* document custom layers cubit

* clean up models code

* disable long press on the map

* clean up map code

* upgrade flutter svg

* clean up empty map

* remove unused import

* fix lint issue

* adding custom layers persistence
