+++ lib/models/custom_layer.dart
@@ -1,0 +1,15 @@
+
+import 'package:flutter_map/flutter_map.dart';
+
+
+abstract class CustomLayer {
+  final String id;
+  Function onRefresh;
+  CustomLayer(this.id);
+  
+  void refresh() {
+    if (onRefresh != null) onRefresh();
+  }
+
+  LayerOptions get layerOptions;
+}
