+++ lib/pages/home/plan_map/plan_map.dart
@@ -1,23 +1,20 @@
+import 'dart:developer';
+
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_map/flutter_map.dart';
-import 'package:geolocator/geolocator.dart';
 import 'package:latlong/latlong.dart';
-import 'package:trufi_core/blocs/gps_location/location_provider_cubit.dart';
 import 'package:trufi_core/entities/plan_entity/plan_entity.dart';
-import 'package:trufi_core/widgets/alerts.dart';
+import 'package:trufi_core/widgets/map/buttons/crop_button.dart';
+import 'package:trufi_core/widgets/map/buttons/map_type_button.dart';
+import 'package:trufi_core/widgets/map/buttons/your_location_button.dart';
+import 'package:trufi_core/widgets/map/map_copyright.dart';
+import 'package:trufi_core/widgets/map/trufi_map_controller.dart';
+import 'package:trufi_core/widgets/map/trufi_map.dart';
 
-import '../composite_subscription.dart';
-import '../plan/plan.dart';
-import '../trufi_app.dart';
-import '../trufi_configuration.dart';
-import '../trufi_map_utils.dart';
-import '../widgets/crop_button.dart';
-import '../widgets/map_type_button.dart';
-import '../widgets/trufi_map.dart';
-import '../widgets/trufi_online_map.dart';
-import '../widgets/your_location_button.dart';
-
+import '../../../composite_subscription.dart';
+import '../../../trufi_app.dart';
+import '../../../widgets/map/utils/trufi_map_utils.dart';
+import './plan.dart';
 const double customOverlayWidgetMargin = 80.0;
 
 typedef OnSelected = void Function(PlanItinerary itinerary);
@@ -55,6 +52,7 @@ class PlanMapPageState extends State<PlanMapPage>
     );
     _subscriptions.add(
       _trufiMapController.outMapReady.listen((_) {
+        log("outMapReady");
         setState(() {
           _data.needsCameraUpdate = true;
         });
@@ -81,7 +79,6 @@ class PlanMapPageState extends State<PlanMapPage>
 
   @override
   Widget build(BuildContext context) {
-    final cfg = TrufiConfiguration();
     final Locale locale = Localizations.localeOf(context);
     final theme = Theme.of(context);
     _data._selectedColor = theme.accentColor;
@@ -97,7 +94,7 @@ class PlanMapPageState extends State<PlanMapPage>
     }
     return Stack(
       children: <Widget>[
-        TrufiOnlineMap(
+        TrufiMap(
           key: const ValueKey("PlanMap"),
           controller: _trufiMapController,
           onTap: _handleOnMapTap,
@@ -109,21 +106,35 @@ class PlanMapPageState extends State<PlanMapPage>
               _data.fromMarkerLayer,
               _data.selectedPolylinesLayer,
               _data.selectedMarkersLayer,
-              _trufiMapController.yourLocationLayer,
+              // _trufiMapController.yourLocationLayer,
               _data.toMarkerLayer,
             ];
           },
         ),
-        if (cfg.map.satelliteMapTypeEnabled || cfg.map.terrainMapTypeEnabled)
-          Positioned(
-            top: 16.0,
-            right: 16.0,
-            child: _buildUpperActionButtons(context),
-          ),
+        const Positioned(
+          top: 16.0,
+          right: 16.0,
+          child: MapTypeButton(),
+        ),
+        Positioned(
+          bottom: 0,
+          left: 0,
+          child: MapCopyright(),
+        ),
         Positioned(
           bottom: 16.0,
           right: 16.0,
-          child: _buildLowerActionButtons(context),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.end,
+            mainAxisAlignment: MainAxisAlignment.end,
+            children: <Widget>[
+              CropButton(key: _cropButtonKey, onPressed: _handleOnCropPressed),
+              const Padding(padding: EdgeInsets.all(4.0)),
+              YourLocationButton(
+                trufiMapController: _trufiMapController,
+              ),
+            ],
+          ),
         ),
         Positioned.fill(
           child: Container(
@@ -153,26 +164,6 @@ class PlanMapPageState extends State<PlanMapPage>
     );
   }
 
-  Widget _buildUpperActionButtons(BuildContext context) {
-    return const SafeArea(
-      child: MapTypeButton(),
-    );
-  }
-
-  Widget _buildLowerActionButtons(BuildContext context) {
-    return SafeArea(
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.end,
-        mainAxisAlignment: MainAxisAlignment.end,
-        children: <Widget>[
-          CropButton(key: _cropButtonKey, onPressed: _handleOnCropPressed),
-          const Padding(padding: EdgeInsets.all(4.0)),
-          YourLocationButton(onPressed: _handleOnYourLocationPressed),
-        ],
-      ),
-    );
-  }
-
   void _handleOnMapTap(LatLng point) {
     final PlanItinerary tappedItinerary = _data.itineraryForPoint(point);
     if (tappedItinerary != null) {
@@ -187,23 +178,6 @@ class PlanMapPageState extends State<PlanMapPage>
     if (_data.selectedBounds != null && _data.selectedBounds.isValid) {
       _cropButtonKey.currentState.setVisible(
         visible: !position.bounds.containsBounds(_data.selectedBounds),
-      );
-    }
-  }
-
-  Future<void> _handleOnYourLocationPressed() async {
-    try {
-      final location =
-          context.read<LocationProviderCubit>().state.currentLocation;
-      _trufiMapController.moveToYourLocation(
-        location: location,
-        context: context,
-        tickerProvider: this,
-      );
-    } on PermissionDeniedException catch (_) {
-      showDialog(
-        context: context,
-        builder: (context) => buildAlertLocationServicesDenied(context),
       );
     }
   }
