+++ pubspec.yaml
@@ -21,7 +21,7 @@ dependencies:
   flutter_localizations:
     sdk: flutter
   flutter_map: ^0.12.0
-  flutter_svg: ^0.21.0+1
+  flutter_svg: ^0.22.0
   geolocator: ^7.0.1
   google_api_availability: ^3.0.1
   http: 0.13.1
